renamed auth to token
