+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/token/CredentialTokenAuth.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.otp.auth;
+package org.aktin.broker.auth.otp.token;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
